Modularizadop del NAV y colocacion de Routes en vez de switch
