+++ client/src/App.js
@@ -1,10 +1,15 @@
-import './App.css';
+import {Routes, Route} from "react-router-dom";
 import Home from './components/Home/Home';
+import Products from './components/Products/Products';
+import './App.css';
 
 function App() {
   return (
     <div className="App">
-      <Home  />
+      <Routes>
+          <Route path = "/" element = {<Home/>}  />
+          <Route exact path = "/productos" element = {<Products />}/>
+      </Routes>
     </div>
   );
 }
