+++ client/src/components/Vcard/Vcard.jsx
@@ -5,7 +5,7 @@ export default function Vcard(props) {
     <div className="card">
        <h4 /*className="flex-items"*/>{props.name}</h4>
        {props.image?<a href={`/videogames/${props.id}`}>
-      <img  className={props.id==1?'imageVcardAlberto':"imageVcard"} src={props.image}  alt="centered image"/>  </a>:null}
+      <img  className={props.id===1?'imageVcardAlberto':"imageVcard"} src={props.image}  alt=""/>  </a>:null}
        
     {/* <h5><strong>{props.image?'Generos':null }</strong> </h5> */}
     {/* <p className="genres">{props.genres.map((object) => ' '+object.name).toString()  }</p> */}
